ELECTRON-1170 - Remove ncp module
