+++ js/main.js
@@ -7,8 +7,6 @@ const crashReporter = electron.crashReporter;
 const nodeURL = require('url');
 const shellPath = require('shell-path');
 const urlParser = require('url');
-const { ncp } = require('ncp');
-const path = require('path');
 const nodePath = require('path');
 require('electron-dl')();
 
@@ -31,8 +29,8 @@ const getCmdLineArg = require('./utils/getCmdLineArg.js');
 
 //setting the env path child_process issue https://github.com/electron/electron/issues/7688
 shellPath()
-    .then((sPath) => {
-        process.env.PATH = sPath
+    .then((path) => {
+        process.env.PATH = path
     })
     .catch(() => {
         process.env.PATH = [
@@ -351,7 +349,6 @@ function setupFirstTimeLaunch(resolve, reject, shouldUpdateUserConfig) {
     log.send(logLevels.INFO, 'setting first time launch config');
     getConfigField('launchOnStartup')
         .then(setStartup)
-        .then(copyDictionaries)
         .then(() => {
             if (shouldUpdateUserConfig) {
                 log.send(logLevels.INFO, `Resetting user config data? ${shouldUpdateUserConfig}`);
@@ -360,37 +357,6 @@ function setupFirstTimeLaunch(resolve, reject, shouldUpdateUserConfig) {
             return resolve();
         })
         .catch(reject);
-}
-
-/**
- * Copies all the dictionaries files required
- * for spellchecker module
- * @return {Promise<any>}
- */
-function copyDictionaries() {
-    return new Promise(resolve => {
-        if (isMac) {
-            resolve();
-            return;
-        }
-        const dictionariesDirName = 'dictionaries';
-        let dictionaryPath;
-        if (isDevEnv) {
-            dictionaryPath = path.join(app.getAppPath(), dictionariesDirName);
-        } else {
-            let execPath = path.dirname(app.getPath('exe'));
-            dictionaryPath = path.join(execPath, isMac ? '..' : '', dictionariesDirName);
-        }
-        const userDictionaries = path.join(app.getPath('userData'), dictionariesDirName);
-
-        ncp(dictionaryPath, userDictionaries, (error) => {
-            if (error) {
-                log.send(logLevels.ERROR, `unable to copy dictionaries ${error}`);
-            }
-            return resolve();
-        });
-
-    })
 }
 
 /**
