+++ package.json
@@ -101,6 +101,7 @@
     "glob": "7.1.3",
     "jest": "23.6.0",
     "jest-html-reporter": "2.4.2",
+    "ncp": "2.0.0",
     "robotjs": "0.5.1",
     "selenium-webdriver": "3.6.0",
     "wdio-selenium-standalone-service": "0.0.12"
@@ -122,7 +123,6 @@
     "lodash.isequal": "4.5.0",
     "lodash.omit": "4.5.0",
     "lodash.pick": "4.4.0",
-    "ncp": "2.0.0",
     "node-osascript": "2.1.0",
     "ref": "1.3.5",
     "shell-path": "2.1.0",
